backend/requirements.txt
asgiref==3.11.0
Django==5.2.9
djangorestframework==3.16.1
pillow==12.0.0
sqlparse==0.5.4
django-cors-headers==4.3.1
djangorestframework-simplejwt
django-filter
gunicorn
whitenoise>=6.6
dj-database-url>=2.2
psycopg2-binary>=2.9